Create solution.php
This solution uses a greedy algorithm to determine the number of packs needed for each pack size, starting with the largest pack size and working downwards. It keeps track of the number of packs needed for each pack size, as well as the total number of packs. Finally, it formats the output string and returns it.